requirements.txt
openai
python-dotenv
requests
yt-dlp
whisper
streamlit
